Added a service for retrieving current user role
