+++ src/test/groovy/vincool/RoleUserServiceSpec.groovy
@@ -1,0 +1,22 @@
+package vincool
+
+import grails.test.mixin.TestFor
+import spock.lang.Specification
+
+/**
+ * See the API for {@link grails.test.mixin.services.ServiceUnitTestMixin} for usage instructions
+ */
+@TestFor(RoleUserService)
+class RoleUserServiceSpec extends Specification {
+
+    def setup() {
+    }
+
+    def cleanup() {
+    }
+
+    void "test something"() {
+        expect:"fix me"
+            true == true
+    }
+}
